+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * AkoBadge, a widget for displaying a list of badges
 * AkoBadgeLabel, a class used as input to AkoBadge
 * AkoBadgeStyle, a class for styling AkoBadge
+* AkoPair, a simple Pair class (2 element Tuple)
 ### Changed
 ### Deprecated
 ### Removed